# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CashFlow Manager API
  version: 0.1.0
  description: Production sync API contract for the offline-first CashFlow Manager Flutter app.
servers:
  - url: http://localhost:3000
    description: Local Docker API
security:
  - bearerAuth: []
paths:
...
  /v1/wallets:
    get:
      summary: List wallets
      responses:
        '200':
          description: Wallet list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
    post:
      summary: Create a wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletInput'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/wallets/{id}:
    patch:
      summary: Update a wallet
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletPatchInput'
      responses:
        '200':
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Soft-delete a wallet
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Wallet tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/categories:
    get:
      summary: List categories
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/categories/{id}:
    patch:
      summary: Update a category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPatchInput'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Soft-delete a category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Category tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/transactions:
    get:
      summary: List transactions
      parameters:
        - name: month
          in: query
          schema:
            type: string
            format: date
          description: First day of the month to filter by.
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FinanceTransaction'
    post:
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/transactions/{id}:
    patch:
      summary: Update a transaction
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionPatchInput'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Soft-delete a transaction
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Transaction tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/budgets:
    get:
      summary: List budgets
      responses:
        '200':
          description: Budget list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
    post:
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetInput'
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/budgets/{id}:
    patch:
      summary: Update a budget
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetPatchInput'
      responses:
        '200':
          description: Budget updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Soft-delete a budget
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Budget tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/saving-goals:
    get:
      summary: List saving goals
      responses:
        '200':
          description: Saving goal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingGoal'
    post:
      summary: Create a saving goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingGoalInput'
      responses:
        '201':
          description: Saving goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingGoal'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/saving-goals/{id}:
    patch:
      summary: Update a saving goal
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingGoalPatchInput'
      responses:
        '200':
          description: Saving goal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingGoal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Soft-delete a saving goal
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Saving goal tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingGoal'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [invalid_amount]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    SyncMetadata:
      type: object
      required: [id, createdAt, updatedAt, revision]
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, 'null']
          format: date-time
        revision:
          type: integer
          minimum: 1
    WalletType:
      type: string
      enum: [cash, bank, eWallet, creditCard]
    TransactionType:
      type: string
      enum: [income, expense, transfer]
    Wallet:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/WalletInput'
    WalletInput:
      type: object
      required: [name, type, initialBalance]
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/WalletType'
        initialBalance:
          type: integer
          minimum: 0
    WalletPatchInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/WalletType'
        initialBalance:
          type: integer
          minimum: 0
    Category:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/CategoryInput'
    CategoryInput:
      type: object
      required: [name, type, colorHex]
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/TransactionType'
        colorHex:
          type: integer
    CategoryPatchInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/TransactionType'
        colorHex:
          type: integer
    FinanceTransaction:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/TransactionInput'
    TransactionInput:
      type: object
      required: [walletId, categoryId, type, amount, date, note, isRecurring]
      properties:
        walletId:
          type: string
          format: uuid
        toWalletId:
          type: [string, 'null']
          format: uuid
        categoryId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: integer
          minimum: 1
        date:
          type: string
          format: date-time
        note:
          type: string
        isRecurring:
          type: boolean
    TransactionPatchInput:
      type: object
      minProperties: 1
      properties:
        walletId:
          type: string
          format: uuid
        toWalletId:
          type: [string, 'null']
          format: uuid
        categoryId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: integer
          minimum: 1
        date:
          type: string
          format: date-time
        note:
          type: string
        isRecurring:
          type: boolean
    Budget:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/BudgetInput'
    BudgetInput:
      type: object
      required: [categoryId, month, limitAmount]
      properties:
        categoryId:
          type: string
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: integer
          minimum: 1
    BudgetPatchInput:
      type: object
      minProperties: 1
      properties:
        categoryId:
          type: string
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: integer
          minimum: 1
    SavingGoal:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/SavingGoalInput'
    SavingGoalInput:
      type: object
      required: [name, targetAmount, savedAmount, deadline]
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: integer
          minimum: 1
        savedAmount:
          type: integer
          minimum: 0
        deadline:
          type: string
          format: date
    SavingGoalPatchInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: integer
          minimum: 1
        savedAmount:
          type: integer
          minimum: 0
        deadline:
          type: string
          format: date
...